project/src/books/requirements.txt
flask
psycopg2-binary
yandex-pgmigrate
prometheus_flask_exporter
prometheus_client
gunicorn
python-keycloak
pika
minio